バイナリセーブ・ロード対応
・MMLパーサを別ファイルに
・バイナリシリアライズ機能を追加
・音色データもstringで保持するようにした
